+++ src/main/java/kitchenpos/menu/domain/Menu.java
@@ -8,7 +8,6 @@ import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import kitchenpos.menu.application.MenuValidator;
 
 @Entity
 public class Menu {
