+++ src/main/java/kitchenpos/menu/domain/MenuValidator.java
@@ -1,4 +1,4 @@
-package kitchenpos.menu.application;
+package kitchenpos.menu.domain;
 
 import java.math.BigDecimal;
 import java.util.List;
