+++ src/main/java/kitchenpos/order/application/OrderService.java
@@ -7,6 +7,7 @@ import kitchenpos.order.domain.OrderRepository;
 import kitchenpos.order.domain.OrderLineItemRepository;
 import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderStatus;
+import kitchenpos.order.domain.OrderValidator;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
