+++ src/main/java/kitchenpos/order/domain/Order.java
@@ -13,7 +13,6 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.OneToMany;
 import javax.persistence.Table;
-import kitchenpos.order.application.OrderValidator;
 
 @Entity
 @Table(name = "orders")
